1. GUI - Using 'Panel', 'Grid', and 'Button' controls of Avalonia, to practice GUI. 2. Method - Implemented a function to get current date time in given format.
[Comment]
GUI coding reminds me when I used to use Swing in Java. This is not hard coded and using XAML instead, so at least it feels better to separate the view and the model parts. I still have choices like WPF or WinForms, but I'd rather stick to Avalonia to figure out its versatility.

[Github]
Resolved - #2